+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/ComparisonExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expression that compares both its contained expressions.
