+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/ConstantExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * An {@link EvaluableExpression} having a constant value.
