+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/DivisionExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expression that divides its contained dividend through its contained divisor.
