+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/EvaluableExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expressions describing relations between {@link EvaluableVariable}s and a
