+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/ExponentationExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expression that potentises the given exponent to the given base.
