+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/ExpressionRecursionException.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Might be thrown if a {@link EvaluableExpression} contains itself.
