+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/IfThenElseExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expression that executes an if-then-else-statement based on its contained expressions.
