+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/LogarithmExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expression that executes a logarithm with defined expressions as base and
