+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/MultiplicationExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 import java.util.Collection;
