+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/NaturalLogarithmExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expression that executes the natural logarithmic function. The antilogarithm is a
