+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/SineExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expression that executes a sine function on its contained expression.
