+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/SubstractionExpression.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
 
 /**
  * Expression that subtracts a subtrahend from a minuend.
