+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/evaluableexpressions/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains evaluable expressions suitable to compose functions.
  */
-package de.uka.ipd.sdq.beagle.core.expressions;
+package de.uka.ipd.sdq.beagle.core.evaluableexpressions;
